refactor: admin sellers dynamic segment [id]를 [applicationId]로 rename
