+++ src/app/api/admin/sellers/[applicationId]/approve/route.test.ts
@@ -7,7 +7,7 @@ import { requireAdmin } from '@/app/api/_lib/auth';
 import { isApiMockEnabled } from '@/app/api/_lib/mock';
 
 import { POST } from './route';
-import { rejectSellerApplication } from '../../_lib/service';
+import { approveSellerApplication } from '../../_lib/service';
 
 vi.mock('@/app/api/_lib/auth', () => ({
   requireAdmin: vi.fn(),
@@ -18,42 +18,34 @@ vi.mock('@/app/api/_lib/mock', () => ({
 }));
 
 vi.mock('../../_lib/service', () => ({
-  rejectSellerApplication: vi.fn(),
+  approveSellerApplication: vi.fn(),
 }));
 
-const SELLER_ID = '00000000-0000-4000-8000-000000000001';
-const REJECT_REASON = '서류 미제출';
+const APPLICATION_ID = '00000000-0000-4000-8000-000000000001';
 
 const adminResult = {
   authUser: {},
   serviceUser: {},
 } as Awaited<ReturnType<typeof requireAdmin>>;
 
-function makePostRequest(body: unknown): NextRequest {
+function makeRequest(): NextRequest {
   return new NextRequest(
-    `http://localhost/api/admin/sellers/${SELLER_ID}/reject`,
-    {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(body),
-    }
+    `http://localhost/api/admin/sellers/${APPLICATION_ID}/approve`,
+    { method: 'POST' }
   );
 }
 
-function makeParams(id = SELLER_ID) {
-  return { params: Promise.resolve({ id }) };
+function makeParams(applicationId = APPLICATION_ID) {
+  return { params: Promise.resolve({ applicationId }) };
 }
 
-describe('POST /api/admin/sellers/[id]/reject', () => {
+describe('POST /api/admin/sellers/[applicationId]/approve', () => {
   beforeEach(() => vi.clearAllMocks());
 
   it('mock 모드에서 null을 반환하고 requireAdmin을 호출하지 않는다', async () => {
     vi.mocked(isApiMockEnabled).mockReturnValue(true);
 
-    const res = await POST(
-      makePostRequest({ reason: REJECT_REASON }),
-      makeParams()
-    );
+    const res = await POST(makeRequest(), makeParams());
     const body = (await res.json()) as { data: null };
 
     expect(res.status).toBe(200);
@@ -61,48 +53,28 @@ describe('POST /api/admin/sellers/[id]/reject', () => {
     expect(requireAdmin).not.toHaveBeenCalled();
   });
 
-  it('real 모드에서 reason 없는 body는 400을 반환한다', async () => {
+  it('real 모드에서 requireAdmin 호출 후 approveSellerApplication을 호출한다', async () => {
     vi.mocked(isApiMockEnabled).mockReturnValue(false);
     vi.mocked(requireAdmin).mockResolvedValue(adminResult);
+    vi.mocked(approveSellerApplication).mockResolvedValue(undefined);
 
-    const res = await POST(makePostRequest({}), makeParams());
-    const body = (await res.json()) as { error: { code: string } };
-
-    expect(res.status).toBe(400);
-    expect(body.error.code).toBe('VALIDATION_ERROR');
-  });
-
-  it('real 모드에서 requireAdmin 호출 후 rejectSellerApplication을 호출한다', async () => {
-    vi.mocked(isApiMockEnabled).mockReturnValue(false);
-    vi.mocked(requireAdmin).mockResolvedValue(adminResult);
-    vi.mocked(rejectSellerApplication).mockResolvedValue(undefined);
-
-    const res = await POST(
-      makePostRequest({ reason: REJECT_REASON }),
-      makeParams()
-    );
+    const res = await POST(makeRequest(), makeParams());
 
     expect(res.status).toBe(200);
     expect(requireAdmin).toHaveBeenCalledOnce();
-    expect(rejectSellerApplication).toHaveBeenCalledWith(
-      SELLER_ID,
-      REJECT_REASON
-    );
+    expect(approveSellerApplication).toHaveBeenCalledWith(APPLICATION_ID);
   });
 
   it('잘못된 UUID params는 400을 반환하고 service를 호출하지 않는다', async () => {
     vi.mocked(isApiMockEnabled).mockReturnValue(false);
     vi.mocked(requireAdmin).mockResolvedValue(adminResult);
 
-    const res = await POST(
-      makePostRequest({ reason: REJECT_REASON }),
-      makeParams('not-a-uuid')
-    );
+    const res = await POST(makeRequest(), makeParams('not-a-uuid'));
     const body = (await res.json()) as { error: { code: string } };
 
     expect(res.status).toBe(400);
     expect(body.error.code).toBe('VALIDATION_ERROR');
-    expect(rejectSellerApplication).not.toHaveBeenCalled();
+    expect(approveSellerApplication).not.toHaveBeenCalled();
   });
 
   it('requireAdmin이 실패하면 error envelope를 반환한다', async () => {
@@ -111,10 +83,7 @@ describe('POST /api/admin/sellers/[id]/reject', () => {
       new AppError(ERROR_CODE.FORBIDDEN, 403)
     );
 
-    const res = await POST(
-      makePostRequest({ reason: REJECT_REASON }),
-      makeParams()
-    );
+    const res = await POST(makeRequest(), makeParams());
     const body = (await res.json()) as { error: { code: string } };
 
     expect(res.status).toBe(403);
