+++ src/app/api/admin/sellers/[applicationId]/reject/route.test.ts
@@ -7,7 +7,7 @@ import { requireAdmin } from '@/app/api/_lib/auth';
 import { isApiMockEnabled } from '@/app/api/_lib/mock';
 
 import { POST } from './route';
-import { approveSellerApplication } from '../../_lib/service';
+import { rejectSellerApplication } from '../../_lib/service';
 
 vi.mock('@/app/api/_lib/auth', () => ({
   requireAdmin: vi.fn(),
@@ -18,34 +18,42 @@ vi.mock('@/app/api/_lib/mock', () => ({
 }));
 
 vi.mock('../../_lib/service', () => ({
-  approveSellerApplication: vi.fn(),
+  rejectSellerApplication: vi.fn(),
 }));
 
-const SELLER_ID = '00000000-0000-4000-8000-000000000001';
+const APPLICATION_ID = '00000000-0000-4000-8000-000000000001';
+const REJECT_REASON = '서류 미제출';
 
 const adminResult = {
   authUser: {},
   serviceUser: {},
 } as Awaited<ReturnType<typeof requireAdmin>>;
 
-function makeRequest(): NextRequest {
+function makePostRequest(body: unknown): NextRequest {
   return new NextRequest(
-    `http://localhost/api/admin/sellers/${SELLER_ID}/approve`,
-    { method: 'POST' }
+    `http://localhost/api/admin/sellers/${APPLICATION_ID}/reject`,
+    {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify(body),
+    }
   );
 }
 
-function makeParams(id = SELLER_ID) {
-  return { params: Promise.resolve({ id }) };
+function makeParams(applicationId = APPLICATION_ID) {
+  return { params: Promise.resolve({ applicationId }) };
 }
 
-describe('POST /api/admin/sellers/[id]/approve', () => {
+describe('POST /api/admin/sellers/[applicationId]/reject', () => {
   beforeEach(() => vi.clearAllMocks());
 
   it('mock 모드에서 null을 반환하고 requireAdmin을 호출하지 않는다', async () => {
     vi.mocked(isApiMockEnabled).mockReturnValue(true);
 
-    const res = await POST(makeRequest(), makeParams());
+    const res = await POST(
+      makePostRequest({ reason: REJECT_REASON }),
+      makeParams()
+    );
     const body = (await res.json()) as { data: null };
 
     expect(res.status).toBe(200);
@@ -53,28 +61,48 @@ describe('POST /api/admin/sellers/[id]/approve', () => {
     expect(requireAdmin).not.toHaveBeenCalled();
   });
 
-  it('real 모드에서 requireAdmin 호출 후 approveSellerApplication을 호출한다', async () => {
+  it('real 모드에서 reason 없는 body는 400을 반환한다', async () => {
     vi.mocked(isApiMockEnabled).mockReturnValue(false);
     vi.mocked(requireAdmin).mockResolvedValue(adminResult);
-    vi.mocked(approveSellerApplication).mockResolvedValue(undefined);
 
-    const res = await POST(makeRequest(), makeParams());
+    const res = await POST(makePostRequest({}), makeParams());
+    const body = (await res.json()) as { error: { code: string } };
+
+    expect(res.status).toBe(400);
+    expect(body.error.code).toBe('VALIDATION_ERROR');
+  });
+
+  it('real 모드에서 requireAdmin 호출 후 rejectSellerApplication을 호출한다', async () => {
+    vi.mocked(isApiMockEnabled).mockReturnValue(false);
+    vi.mocked(requireAdmin).mockResolvedValue(adminResult);
+    vi.mocked(rejectSellerApplication).mockResolvedValue(undefined);
+
+    const res = await POST(
+      makePostRequest({ reason: REJECT_REASON }),
+      makeParams()
+    );
 
     expect(res.status).toBe(200);
     expect(requireAdmin).toHaveBeenCalledOnce();
-    expect(approveSellerApplication).toHaveBeenCalledWith(SELLER_ID);
+    expect(rejectSellerApplication).toHaveBeenCalledWith(
+      APPLICATION_ID,
+      REJECT_REASON
+    );
   });
 
   it('잘못된 UUID params는 400을 반환하고 service를 호출하지 않는다', async () => {
     vi.mocked(isApiMockEnabled).mockReturnValue(false);
     vi.mocked(requireAdmin).mockResolvedValue(adminResult);
 
-    const res = await POST(makeRequest(), makeParams('not-a-uuid'));
+    const res = await POST(
+      makePostRequest({ reason: REJECT_REASON }),
+      makeParams('not-a-uuid')
+    );
     const body = (await res.json()) as { error: { code: string } };
 
     expect(res.status).toBe(400);
     expect(body.error.code).toBe('VALIDATION_ERROR');
-    expect(approveSellerApplication).not.toHaveBeenCalled();
+    expect(rejectSellerApplication).not.toHaveBeenCalled();
   });
 
   it('requireAdmin이 실패하면 error envelope를 반환한다', async () => {
@@ -83,7 +111,10 @@ describe('POST /api/admin/sellers/[id]/approve', () => {
       new AppError(ERROR_CODE.FORBIDDEN, 403)
     );
 
-    const res = await POST(makeRequest(), makeParams());
+    const res = await POST(
+      makePostRequest({ reason: REJECT_REASON }),
+      makeParams()
+    );
     const body = (await res.json()) as { error: { code: string } };
 
     expect(res.status).toBe(403);
